requirements.txt
streamlit
opencv-python
pandas
numpy
paho-mqtt
Pillow
ultralytics
cvzone
